requirements.txt
fastapi>=0.115.12,<0.116.0
sqlalchemy>=2.0.39,<3.0.0
pydantic>=2.10.6,<3.0.0
alembic>=1.15.1,<2.0.0
uvicorn>=0.34.0,<0.35.0
python-dotenv>=1.1.0,<2.0.0
asyncpg>=0.29.0,<1.0.0
